refactor: fix multi-module compilation and update CI/Docker paths
- Change internal→public for cross-module test access (ComicPageHashService,
  FilenameMetadataService, SidecarMetadataService)
- Explicit cast for smart cast across module boundary in CombinedFeaturesIntegrationTest
- Remove 17 duplicate dependencies from runary-web/pom.xml (come transitively)
- Update Dockerfile for multi-module source layout and fat JAR path
- Update all CI workflows to use **/target/ globs and runary-web/target/ paths
- Fix .gitignore icon path for new module structure

+++ runary-web/src/test/kotlin/org/runary/integration/CombinedFeaturesIntegrationTest.kt
@@ -257,7 +257,7 @@ class CombinedFeaturesIntegrationTest : IntegrationTestBase() {
         val prog = Json.mapper.readValue(r.bodyString(), ReadingProgressDto::class.java)
         // percentage may be null if pageCount not yet extracted; if set it should be 100.0
         if (prog.percentage != null) {
-            assertEquals(100.0, prog.percentage, 0.01)
+            assertEquals(100.0, prog.percentage as Double, 0.01)
         }
     }
